[hotfix] : baseIndex 자료형 수정(Integer > BigDecimal) (#4)
* fix : baseIndex 자료형 수정(Integer > BigDecimal)

* fix : DDL 수정 (INT > NUMERIC)

+++ src/main/java/com/codeit/team7/findex/dto/response/StockMarketIndexResponse.java
@@ -119,7 +119,7 @@ public class StockMarketIndexResponse {
     @JsonFormat(pattern = "yyyyMMdd")
     private LocalDate basePointInDate;
     @JsonProperty("basIdx")
-    private Integer baseIndex;
+    private BigDecimal baseIndex;
   }
 }
